+++ src/app/app.routes.spec.ts
@@ -12,7 +12,9 @@ describe('routes', () => {
     expect(routePaths(routes)).toEqual([
       '',
       PATH_NAMES.BLOG,
+      `${PATH_NAMES.BLOG}/search`,
       `${PATH_NAMES.BLOG}/category/:category`,
+      `${PATH_NAMES.BLOG}/tag/:tag`,
       `${PATH_NAMES.BLOG}/:slug`,
       PATH_NAMES.LABS,
       PATH_NAMES.FS_BACKGROUND,